(FIXME) WebGPU: Implement Futures infrastructure + RequestAdapterF(0)
FIXME: This patches in the necessary parts of the C header changes but
doesn't update the C++ header. This should be done with an actual header
roll from Dawn.

Timeouts (which require Asyncify) are not implemented yet, but
I've spot-checked that all three CallbackModes work.

Sadly this adds substantial binary size (probably due to C++ STL).

- Before:
  hello.js:   28316 bytes / 11064 gzipped
  hello.wasm: 24544 bytes / 10445 gzipped
  total:      52860 bytes / 21509 gzipped separately
- After:
  hello.js:   28613 bytes / 11161 gzipped
  hello.wasm: 34010 bytes / 13686 gzipped
  total:      62623 bytes / 24847 gzipped separately